fix satellite slow queries
In this change we continue using the cache for the frame count, but instead of using the cache for the latest frame as discussed in IRC we grab the latest frame based on the ID.  In my test environment this was a reasonable query time.

I'd rather we try this route before resorting to the cache for latest frame as it will be a better user experience.

Signed-off-by: Corey Shields <cshields@gmail.com>