Update 'satellitetle'
In the 0.6.0 release the main CelesTrak source was changed from
'Satellites (active)' to the new 'SatNOGS' list. This list is
composed by CelesTrak by fetching the list of satellites
from satnogs-db.

This doesn't introduce a circular dependency:
As before, if a TLE is not available from the main collections
(which beside the previously discussed CelesTrak list includes
another from AMSAT and CalPoly), the CelesTrak SATCAT API is
questioned for each missing satellite separately.

Signed-off-by: Fabian P. Schmidt <kerel-fs@gmx.de>