Linting for R0914
A couple of functions in tasks and utils were refactored with new helper functions to reduce the number of local variables per function.

In addition, fetch_data.py was flagged with R0914 but this is a relic of the original attempt at decoding data into DB, and is unused - so I am removing rather than fixing it.

Signed-off-by: Corey Shields <cshields@gmail.com>