Revert "Remove eslint.failAfterError()"
This reverts commit a088e5d03b627e78a37c779deca71fc0123e7833.

The commit had been reverted as it was disabling failing the pipeline
in CI on errors. Without the eslint.failAfterError(), the exit status
of gulp was 0.

For solving the issue that the reverted commit was trying to solve,
continue running parallel tasks if one of them fails, gulp needs to
be run with the flag --continue, however note that this flag allows
tasks in series to continue run even if one of them fails.