Refacter transmitter_card fields and kaitai struct decoder info
Removes the field.html include - and the frequent inclusion of that in the transmitter cards.

Fix a bug in satellite.js where input not recognized as an integer

Add gitlab links to the version info

Removes the manually-entered kaitai struct field from view (will need removed from db in a later change) - replacing with a procedurally generated list of fields from the class imported through satnogsdecoders.

Fixes #385

Signed-off-by: Corey Shields <cshields@gmail.com>