Add decoded data recent history chart to satellite view
Adds a chart card showing the last month of decoded data as currently stored in influxdb.

Unfortunately there's no native way to count the number of 'entries' for a measurement across a given time, so we have to do a count(*) across all points. This returns a blob of <timestamp>, <count(pointa)>, <count(pointb)>, <...> and in almost (but not all) cases those counts will be identical but we have to account for it not, so I iterate over each timestamp (client side) and take the max count to assign to that point in time. Since we are doing a 30d query from influx with 1d aggregation this should not be too intensive.

Also cleaned up the coloring of the profile link buttons.

Signed-off-by: Corey Shields <cshields@gmail.com>