Fix export last week bug
As reported in #433 there was a bug causing export requests for the last week to be treated as a last month export.

Fixes #433 by treating the period variable as the int that it is. Also makes the preceding check a little easier to understand by being explicit against the None case. 

Signed-off-by: Corey Shields <cshields@gmail.com>