Add sorting by status to satellites view
Adds the ability to sort on the 'status' column (which today does nothing) by setting the field by which datatables will sort on to sat.status

This also has the benefit of filtering on sat.status as well, though one would need to know the backend statuses to use (ie: 'alive' vs 'dead'). Not sure that is ideal but this is a quick fix.

Fixes #428 

Signed-off-by: Corey Shields <cshields@gmail.com>