bump container from python 3.8.6 to 3.8.7, revert gunicorn
There are a slew of issues related to eventlet monkeypatching ssl.py which leads to the issues we've seen in the updates I pushed to dev this week. The gunicorn update to 20 seems to be the culprit here, and wasn't caught locally because "develop" in djangoctl.sh uses the django runserver, not gunicorn.

Hopefully this fixes our issues, and I'll slowly update eventlet, requests, and urllib3 back up
 
Signed-off-by: Corey Shields <cshields@gmail.com>