Add version field for demoddata entries
Adds an optional 'version' field which will store version details for clients that are sending it (such as some SiDS clients, even though version is not clearly indicated in the protocol pdf)

I set this as a 45 length CharField, anticipating that there may be some application names, versions, and possible git hashes included as well.

Adds a db migration for this field, blank by default.
 
also removed an unnecessary setting which was default (see !628)

Relates to #456

Signed-off-by: Corey Shields <cshields@gmail.com>